+++ R/subscript.R
@@ -64,17 +64,25 @@ vec_as_subscript_result <- function(i,
 #' @export
 vec_as_subscript2 <- function(i,
                               ...,
-                              logical = c("cast", "error"),
                               numeric = c("cast", "error"),
                               character = c("cast", "error"),
                               arg = NULL,
                               call = caller_env()) {
-  check_dots_empty0(...)
+  check_dots <- function(..., logical = "error", call = caller_env()) {
+    if (!is_string(logical, "error")) {
+      abort(
+        "`vctrs::vec_as_subscript2(logical = 'cast')` is deprecated.",
+        call = caller_env()
+      )
+    }
+    check_dots_empty0(..., call = call)
+  }
+  check_dots(...)
+
   result_get(vec_as_subscript2_result(
     i,
     arg,
     call,
-    logical = logical,
     numeric = numeric,
     character = character
   ))
@@ -82,10 +90,8 @@ vec_as_subscript2 <- function(i,
 vec_as_subscript2_result <- function(i,
                                      arg,
                                      call,
-                                     logical = "cast",
                                      numeric = "cast",
                                      character = "cast") {
-  logical <- arg_match0(logical, c("cast", "error"))
   numeric <- arg_match0(numeric, c("cast", "error"))
   character <- arg_match0(character, c("cast", "error"))
 
@@ -93,46 +99,20 @@ vec_as_subscript2_result <- function(i,
     i,
     arg = arg,
     call = call,
-    logical = logical,
+    logical = "error",
     numeric = numeric,
     character = character
   )
 
-  # Return a child of subscript error. The child error messages refer
-  # to single subscripts instead of subscript vectors.
+  # Transform `[` errors into `[[` errors
   if (!is_null(result$err)) {
-    parent <- result$err$parent
-    if (inherits(parent, "vctrs_error_cast_lossy")) {
-      bullets <- new_cnd_bullets_subscript_lossy_cast(parent)
-    } else {
-      bullets <- cnd_body.vctrs_error_subscript_type
-    }
-
     result$err <- new_error_subscript2_type(
       i = result$err$i,
-      logical = logical,
       numeric = numeric,
       character = character,
       subscript_arg = arg,
-      body = bullets,
       call = call
     )
-
-    return(result)
-  }
-
-  i <- result$ok
-
-  if (typeof(i) == "logical") {
-    return(result(err = new_error_subscript2_type(
-      i = i,
-      logical = logical,
-      numeric = numeric,
-      character = character,
-      subscript_arg = arg,
-      body = cnd_body.vctrs_error_subscript_type,
-      call = call
-    )))
   }
 
   result
@@ -254,13 +234,12 @@ cnd_header.vctrs_error_subscript_size <- function(cnd, ...) {
 }
 
 new_error_subscript2_type <- function(i,
-                                      logical,
                                       numeric,
                                       character,
                                       ...) {
   new_error_subscript_type(
     i = i,
-    logical = logical,
+    logical = "error",
     numeric = numeric,
     character = character,
     subscript_scalar = TRUE,